+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Types of changes are:
 * **Fixed** for any bug fixes.
 
 ## [Unreleased]
+
+## [0.1.0] - 2019-12-13
 ### Added
 * Command run as `statham --input {reference} [--output {output}]`
 * Resolve local and remote JSONSchema references to load and
@@ -54,7 +56,8 @@ Types of changes are:
 ## [0.0.0]
 Nothing here.
 
-[Unreleased]: https://github.com/jacksmit15/statham-schema/compare/initial..HEAD
+[Unreleased]: http://github.com/jacksmith15/statham-schema/compare/0.1.0..HEAD
+[0.1.0]: http://github.com/jacksmith15/statham-schema/compare/0.0.0..0.1.0
 
 [Keep a Changelog]: http://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: http://semver.org/spec/v2.0.0.html